+++ app/javascript/main.jsx
@@ -9,7 +9,7 @@ import Signup from "./routes/Signup";
 import SignIn from './components/sessions/SignIn';
 import AuthProvider from './components/sessions/AuthContext';
 import RoomView from "./components/rooms/RoomView";
-import RoomsView from "./components/rooms/RoomsView";
+import Rooms from "./components/rooms/Rooms";
 
 const queryClient = new QueryClient()
 
@@ -22,7 +22,7 @@ const root = (
                    <Route index element={<h1 className="text-center">Index</h1>} />
                    <Route path="/signup" element={<Signup />} />
                    <Route path="/sign_in" element={<SignIn />} />
-                   <Route path="/rooms" element={<RoomsView/>} />
+                   <Route path="/rooms" element={<Rooms/>} />
                    <Route path="/rooms/:friendly_id" element={<RoomView />} />
                    <Route path="*" element={<h1 className="text-center">404</h1>} />
                  </Route>
